backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
httpx==0.28.1
openai==2.14.0
groq==1.0.0
pydantic==2.12.5
python-multipart==0.0.21
beautifulsoup4==4.14.3
lxml==6.0.2
pdfplumber==0.11.8
PyPDF2==3.0.1
python-docx==1.2.0
requests==2.32.5
